fix rpi lights - environmentmap const - require vulkan 1.1 - refactoring

+++ system_modules/naprenderadvanced/src/lightcomponent.cpp
@@ -31,7 +31,6 @@ RTTI_BEGIN_CLASS_NO_DEFAULT_CONSTRUCTOR(nap::LightComponent)
 	RTTI_PROPERTY("Color",				&nap::LightComponent::mColor,				nap::rtti::EPropertyMetaData::Required)
 	RTTI_PROPERTY("Intensity",			&nap::LightComponent::mIntensity,			nap::rtti::EPropertyMetaData::Required)
 	RTTI_PROPERTY("ShadowStrength",		&nap::LightComponent::mShadowStrength,		nap::rtti::EPropertyMetaData::Default)
-	RTTI_PROPERTY("ShadowSampleCount",	&nap::LightComponent::mShadowSampleCount,	nap::rtti::EPropertyMetaData::Default)
 	RTTI_PROPERTY("Enable Shadows",		&nap::LightComponent::mEnableShadows,		nap::rtti::EPropertyMetaData::Default)
 RTTI_END_CLASS
 
@@ -64,7 +63,6 @@ namespace nap
 		mIsEnabled = mResource->mEnabled;
 		mIsShadowEnabled = mResource->mEnableShadows;
 		mShadowStrength = mResource->mShadowStrength;
-		mShadowSampleCount = mResource->mShadowSampleCount;
 
 		// Fetch transform
 		mTransform = &getEntityInstance()->getComponent<TransformComponentInstance>();
@@ -75,16 +73,19 @@ namespace nap
 
 		if (mIsShadowEnabled)
 		{
-			if (!errorState.check(getShadowCamera() != nullptr, "%s: Shadows are enabled No shadow camera set", mID.c_str()))
+			if (!errorState.check(getShadowCamera() != nullptr, "%s: Shadows are enabled while no shadow camera is set", mID.c_str()))
 				return false;
 		}
 
-		// Register with service
-		auto* service = getEntityInstance()->getCore()->getService<RenderAdvancedService>();
-		assert(service != nullptr);
-		service->registerLightComponent(*this);
+        // Register with service
+        if (mIsEnabled)
+        {
+            auto *service = getEntityInstance()->getCore()->getService<RenderAdvancedService>();
+            assert(service != nullptr);
+            service->registerLightComponent(*this);
 
-		mIsRegistered = true;
+            mIsRegistered = true;
+        }
 		return true;
 	}
 
@@ -109,10 +110,13 @@ namespace nap
 
 	void LightComponentInstance::removeLightComponent()
 	{
-		auto* service = getEntityInstance()->getCore()->getService<RenderAdvancedService>();
-		assert(service != nullptr);
-		service->removeLightComponent(*this);
+        if (mIsRegistered)
+        {
+            auto *service = getEntityInstance()->getCore()->getService<RenderAdvancedService>();
+            assert(service != nullptr);
+            service->removeLightComponent(*this);
 
-		mIsRegistered = false;
+            mIsRegistered = false;
+        }
 	}
 }
